pyproject.toml
[project]
name = "decision-intelligence-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.0",
    "groq>=1.2.0",
    "joblib>=1.5.3",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
    "textblob>=0.20.0",
    "uvicorn>=0.45.0",
    "xgboost>=3.2.0",
]

